Veritas/api/requirements.txt
fastapi
uvicorn
python-multipart
imagehash
Pillow
numpy
scipy
algosdk
